+++ src/lib/AsyncTypeaheadInput.tsx
@@ -26,6 +26,7 @@ const AsyncTypeaheadInput = <T extends FieldValues>(props: AsyncTypeaheadProps<T
       control={control}
       name={name}
       render={({ field, fieldState: { error } }) => (
+        // todo: pass in the labelToolTip here and also in all other components that use FormGroupLayout
         <FormGroupLayout helpText={helpText} name={name} id={id} label={label}>
           <AsyncTypeahead
             {...field}