+++ src/lib/types/CommonInputProps.ts
@@ -5,10 +5,12 @@ interface CommonInputProps<T extends FieldValues> {
   onChange?: (e: React.ChangeEvent<HTMLInputElement>) => void;
   onBlur?: (e: React.FocusEvent<HTMLInputElement>) => void;
   label: string;
+  // todo: change this to a ReactNode so that the caller can pass in anything they want
   name: keyof T;
   id?: string;
   helpText?: ReactNode;
   disabled?: boolean;
+  // todo: add labelToolTip here as an optional string property
 }
 
 export { CommonInputProps };